Update version to v0.5.0
With the addition of the initial translatable attributes and I18n support, this warrants bumping the Community Engine to a new minor version.

I18n support is still not complete (there are still a lot of strings that are not using translations), so this will be expanded upon in future updates. However, the I18n features included in this version seem stable enough to be included in the main branch.