Allow navigation items to be assigned dynamic routes used to genereate the URL.
This addresses the issue where internal paths to controller index actions were stored in the db as static paths with a fixed locale (default). Now, routes are able to be dynamically dispatched when generating urls, ensuring that the locale in the URL is the current locale.

Also improved handling/setting linkable id and type, and defer to the linkable to determine the visibility of the nav item if linkable is present.